refactor(megalog): cleanup client event utils
